# packages/backend/src/infrastructure/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Equipment Rental System API
  description: |
    A comprehensive API for managing equipment rentals, reservations, and members.
    Built using hexagonal architecture with TypeScript.

    ## Features
    - Equipment management and availability tracking
    - Member registration and tier-based benefits
    - Rental creation, extension, and return workflows
    - Reservation system with payment authorization
    - Damage assessment and late fee calculation
    - Payment processing integration
    - Notification system

    ## Architecture
    This API follows hexagonal (ports and adapters) architecture principles,
    ensuring clean separation of concerns and testability.
  version: 1.0.0
  contact:
    name: Equipment Rental System
    email: support@equipmentrental.com
servers:
  - url: http://localhost:3000
    description: Development server
  - url: http://localhost:3000/api
    description: API base path
...
paths:
...
  /api/reservations:
    post:
      summary: Create a new reservation
      description: Creates a new reservation for future equipment rental
      tags:
        - Reservations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequest'
            example:
              equipmentId: eq_123
              memberId: mem_789
              startDate: '2024-02-01T00:00:00.000Z'
              endDate: '2024-02-05T00:00:00.000Z'
              paymentMethod:
                type: credit_card
                token: tok_visa_4242
              authorizePayment: true
              notificationChannel: EMAIL
      responses:
        '201':
          description: Reservation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateReservationResponse'
              example:
                reservationId: res_001
                equipmentName: Professional Camera
                startDate: '2024-02-01T00:00:00.000Z'
                endDate: '2024-02-05T00:00:00.000Z'
                authorizationId: auth_123456
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/reservations/{reservationId}:
    get:
      summary: Get reservation by ID
      description: Retrieves detailed information about a specific reservation
      tags:
        - Reservations
      parameters:
        - $ref: '#/components/parameters/ReservationId'
      responses:
        '200':
          description: Reservation details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
              example:
                reservationId: res_001
                equipmentId: eq_123
                memberId: mem_789
                startDate: '2024-02-01T00:00:00.000Z'
                endDate: '2024-02-05T00:00:00.000Z'
                status: PENDING
                createdAt: '2024-01-15T10:00:00.000Z'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/reservations/{reservationId}/cancel:
    put:
      summary: Cancel a reservation
      description: Cancels an existing reservation and processes refunds if applicable
      tags:
        - Reservations
      parameters:
        - $ref: '#/components/parameters/ReservationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelReservationRequest'
            example:
              reason: Changed plans
              authorizationId: auth_123456
              notificationChannel: EMAIL
      responses:
        '200':
          description: Reservation cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelReservationResponse'
              example:
                reservationId: res_001
                cancelledAt: '2024-01-15T11:00:00.000Z'
                refundAmount: 200.00
                refundTransactionId: ref_123456
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/reservations/{reservationId}/confirm:
    put:
      summary: Confirm a reservation
      description: Confirms a pending reservation
      tags:
        - Reservations
      parameters:
        - $ref: '#/components/parameters/ReservationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmReservationRequest'
            example:
              notificationChannel: EMAIL
      responses:
        '200':
          description: Reservation confirmed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmReservationResponse'
              example:
                reservationId: res_001
                confirmedAt: '2024-01-15T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/reservations/{reservationId}/fulfill:
    post:
      summary: Fulfill a reservation
      description: Converts a confirmed reservation into an active rental
      tags:
        - Reservations
      parameters:
        - $ref: '#/components/parameters/ReservationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FulfillReservationRequest'
            example:
              paymentMethod:
                type: credit_card
                token: tok_visa_4242
              authorizationId: auth_123456
              notificationChannel: EMAIL
      responses:
        '200':
          description: Reservation fulfilled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FulfillReservationResponse'
              example:
                reservationId: res_001
                rentalId: rent_002
                totalCost: 200.00
                transactionId: txn_999888
                paymentStatus: succeeded
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  parameters:
...
    ReservationId:
      name: reservationId
      in: path
      required: true
      description: Unique identifier for the reservation
      schema:
        type: string
        example: res_001
  schemas:
...
    CreateReservationRequest:
      type: object
      required:
        - equipmentId
        - memberId
        - startDate
        - endDate
      properties:
        equipmentId:
          type: string
          description: ID of equipment to reserve
        memberId:
          type: string
          description: ID of member creating reservation
        startDate:
          type: string
          format: date-time
          description: Reservation start date (ISO 8601)
        endDate:
          type: string
          format: date-time
          description: Reservation end date (ISO 8601)
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        authorizePayment:
          type: boolean
          description: Whether to authorize payment upfront
          default: false
        notificationChannel:
          type: string
          enum: [EMAIL, SMS]
          description: Preferred notification channel
    CreateReservationResponse:
      type: object
      required:
        - reservationId
        - equipmentName
        - startDate
        - endDate
      properties:
        reservationId:
          type: string
          description: Unique identifier for the reservation
        equipmentName:
          type: string
          description: Name of reserved equipment
        startDate:
          type: string
          format: date-time
          description: Reservation start date
        endDate:
          type: string
          format: date-time
          description: Reservation end date
        authorizationId:
          type: string
          description: Payment authorization ID if applicable
    CancelReservationRequest:
      type: object
      properties:
        reason:
          type: string
          description: Reason for cancellation
        authorizationId:
          type: string
          description: Payment authorization ID to release
        notificationChannel:
          type: string
          enum: [EMAIL, SMS]
          description: Preferred notification channel
    CancelReservationResponse:
      type: object
      required:
        - reservationId
        - cancelledAt
      properties:
        reservationId:
          type: string
          description: ID of cancelled reservation
        cancelledAt:
          type: string
          format: date-time
          description: Timestamp of cancellation
        refundAmount:
          type: number
          format: float
          description: Amount refunded if applicable
        refundTransactionId:
          type: string
          description: Refund transaction ID if applicable
    ConfirmReservationRequest:
      type: object
      properties:
        notificationChannel:
          type: string
          enum: [EMAIL, SMS]
          description: Preferred notification channel
    ConfirmReservationResponse:
      type: object
      required:
        - reservationId
        - confirmedAt
      properties:
        reservationId:
          type: string
          description: ID of confirmed reservation
        confirmedAt:
          type: string
          format: date-time
          description: Timestamp of confirmation
    FulfillReservationRequest:
      type: object
      required:
        - paymentMethod
      properties:
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        authorizationId:
          type: string
          description: Payment authorization ID to capture
        notificationChannel:
          type: string
          enum: [EMAIL, SMS]
          description: Preferred notification channel
    FulfillReservationResponse:
      type: object
      required:
        - reservationId
        - rentalId
        - totalCost
        - transactionId
        - paymentStatus
      properties:
        reservationId:
          type: string
          description: ID of fulfilled reservation
        rentalId:
          type: string
          description: ID of created rental
        totalCost:
          type: number
          format: float
          description: Total rental cost
        transactionId:
          type: string
          description: Payment transaction ID
        paymentStatus:
          type: string
          enum: [succeeded, pending, failed]
          description: Status of payment
    Reservation:
      type: object
      required:
        - reservationId
        - equipmentId
        - memberId
        - startDate
        - endDate
        - status
        - createdAt
      properties:
        reservationId:
          type: string
          description: Unique identifier for the reservation
        equipmentId:
          type: string
          description: ID of reserved equipment
        memberId:
          type: string
          description: ID of member who reserved
        startDate:
          type: string
          format: date-time
          description: Reservation start date
        endDate:
          type: string
          format: date-time
          description: Reservation end date
        status:
          type: string
          enum: [PENDING, CONFIRMED, FULFILLED, CANCELLED, EXPIRED]
          description: Current reservation status
        createdAt:
          type: string
          format: date-time
          description: Timestamp when reservation was created

    # Common Schemas
    PaymentMethod:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          description: Type of payment method
          enum: [credit_card, debit_card, cash]
        token:
          type: string
          description: Payment token from payment provider
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error details
              additionalProperties: true
    ValidationErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR]
              description: Validation error code
            message:
              type: string
              description: Human-readable error message
            fields:
              type: array
              description: List of field validation errors
              items:
                type: object
                required:
                  - field
                  - message
                properties:
                  field:
                    type: string
                    description: Field name with error
                  message:
                    type: string
                    description: Validation error message
  responses:
    BadRequest:
      description: Invalid request parameters or validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            error:
              code: VALIDATION_ERROR
              message: Invalid request parameters
              fields:
                - field: startDate
                  message: Start date must be in the future
                - field: endDate
                  message: End date must be after start date
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            equipmentNotFound:
              value:
                error:
                  code: EQUIPMENT_NOT_FOUND
                  message: Equipment not found
                  details:
                    equipmentId: eq_123
            memberNotFound:
              value:
                error:
                  code: MEMBER_NOT_FOUND
                  message: Member not found
                  details:
                    memberId: mem_789
            rentalNotFound:
              value:
                error:
                  code: RENTAL_NOT_FOUND
                  message: Rental not found
                  details:
                    rentalId: rent_001
    Conflict:
      description: Resource conflict or business rule violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            equipmentNotAvailable:
              value:
                error:
                  code: EQUIPMENT_NOT_AVAILABLE
                  message: Equipment is not available for the requested period
                  details:
                    equipmentId: eq_123
                    startDate: '2024-01-15'
                    endDate: '2024-01-20'
            rentalAlreadyReturned:
              value:
                error:
                  code: RENTAL_ALREADY_RETURNED
                  message: Rental has already been returned
                  details:
                    rentalId: rent_001
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INTERNAL_SERVER_ERROR
              message: An unexpected error occurred